requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
supabase==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
python-dotenv==1.0.1 